Split off example implementations into `examples`. Also created an example PhaseConstraint.
